+++ src/main/java/com/bejker/interactionmanager/gui/BlockBlacklistScreen.java
@@ -14,63 +14,26 @@ import java.util.Locale;
 import java.util.Optional;
 
 
-public class BlockBlacklistScreen extends GameOptionsScreen {
+public class BlockBlacklistScreen extends BlacklistScreen {
 
     private static final Text TITLE_TEXT = Text.translatable("screen.interactionmanager.block_blacklist");
 
-    private final Screen parent;
     private BlockListWidget blockList;
 
-    private TextWidget titleWidget;
-    private TextFieldWidget search;
-
     public BlockBlacklistScreen(Screen parent) {
-        super(parent,MinecraftClient.getInstance().options, TITLE_TEXT);
-        this.parent = parent;
-    }
-
-    @Override
-    protected void initHeader() {
-        DirectionalLayoutWidget widget = new DirectionalLayoutWidget(0,0, DirectionalLayoutWidget.DisplayAxis.VERTICAL);
-        widget.getMainPositioner().margin(0,5);
-        widget.getMainPositioner().alignVerticalCenter();
-        widget.setY(0);
-        titleWidget = new TextWidget(title,textRenderer);
-        titleWidget.alignCenter();
-        widget.add(titleWidget);
-        search = new TextFieldWidget(textRenderer,150,20,Text.empty());
-        //search.setX(width / 2 - search.getWidth() / 2);
-        widget.add(search);
-        this.layout.addHeader(widget);
-        this.layout.setHeaderHeight(this.layout.getHeaderHeight() + 20);
+        super(parent, TITLE_TEXT);
     }
 
     @Override
     protected void initBody() {
-        this.addDrawableChild(search);
+        super.initBody();
         blockList = layout.addBody(new BlockListWidget(this, this.client));
         this.refreshWidgetPositions();
     }
 
     protected void refreshWidgetPositions() {
-        this.layout.refreshPositions();
+        super.refreshWidgetPositions();
         this.blockList.position(this.width, this.layout);
-        titleWidget.setX(titleWidget.getX() + titleWidget.getWidth() / 2);
     }
 
-    @Override
-    protected void addOptions() {
-    }
-
-    @Override
-    public void close() {
-        if(client == null){
-            return;
-        }
-        client.setScreen(parent);
-    }
-
-    public String getSearch() {
-        return this.search.getText().strip().toLowerCase(Locale.ROOT);
-    }
 }