+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+## 1.3.0
+### Added
+- Entity blacklist
+- Searching by block id (e.g. minecraft:stone)
+### Changed
+- Block blacklist button, and entity blacklist button are visually disabled if they are disabled
+## 1.2.0
+### Added 
+- A button to toggle the block blacklist
+### Fixed
+- Game crash when searching, and the language is set to one using non-english characters
 ## 1.1.2
 ### Fixed
 - Game crash when loading an older config